feat: remove old files
